+++ include/parsing.h
@@ -6,7 +6,7 @@
 /*   By: hsarhan <hsarhan@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/17 13:45:41 by hsarhan           #+#    #+#             */
-/*   Updated: 2023/01/24 18:15:01 by hsarhan          ###   ########.fr       */
+/*   Updated: 2023/01/26 20:09:30 by hsarhan          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -221,7 +221,7 @@ bool	print_cone_error(t_shape_errors *err, const char *line, int line_num);
 bool	is_settings(const char *line);
 bool	is_valid_key(const char *key);
 bool	is_valid_color(const char *color);
-bool	check_value(const char *key, const char *val, double min, double max);
+bool	check_value(const char *key, const char *val, float min, float max);
 bool	is_valid_val(const char *key, const char *val);
 t_color	parse_color_value(const char *str);
 bool	parse_setting(t_shape *shape, char **key_val);
