+++ src/edit_mode/loop_hook.c
@@ -6,7 +6,7 @@
 /*   By: hsarhan <hsarhan@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/20 18:50:31 by hsarhan           #+#    #+#             */
-/*   Updated: 2023/01/23 12:42:27 by hsarhan          ###   ########.fr       */
+/*   Updated: 2023/01/26 20:09:30 by hsarhan          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	handle_color_change(int key, t_scene *scene, t_color *color);
 
-t_vector	*sphere_to_xyz(t_vector *vec, double phi, double theta, double r)
+t_vector	*sphere_to_xyz(t_vector *vec, float phi, float theta, float r)
 {
 	vec->x = r * sin(phi) * cos(theta);
 	vec->z = r * sin(phi) * sin(theta);
@@ -115,7 +115,7 @@ void	move_object_v(t_scene *scene, t_shape *shape)
 
 void	scale_object(t_scene *scene, t_shape *shape)
 {
-	double	amount;
+	float	amount;
 
 	amount = 0.04;
 	if (scene->keys_held.minus == true)
@@ -155,7 +155,7 @@ void	change_height(t_scene *scene, t_shape *shape)
 	}
 }
 
-void	rotate_x(t_scene *scene, t_mat4 *rot_mat, double deg)
+void	rotate_x(t_scene *scene, t_mat4 *rot_mat, float deg)
 {
 	t_mat4		rot;
 	t_vector	ax;
@@ -176,7 +176,7 @@ void	rotate_x(t_scene *scene, t_mat4 *rot_mat, double deg)
 	mat_multiply(rot_mat, &rot, &mat_copy);
 }
 
-void	rotate_y(t_scene *scene, t_mat4 *rot_mat, double deg)
+void	rotate_y(t_scene *scene, t_mat4 *rot_mat, float deg)
 {
 	t_mat4	rot;
 	t_mat4	mat_copy;
@@ -189,7 +189,7 @@ void	rotate_y(t_scene *scene, t_mat4 *rot_mat, double deg)
 	mat_multiply(rot_mat, &rot, &mat_copy);
 }
 
-void	rotate_z(t_scene *scene, t_mat4 *rot_mat, double deg)
+void	rotate_z(t_scene *scene, t_mat4 *rot_mat, float deg)
 {
 	t_mat4	rot;
 	t_mat4	mat_copy;
